Added a mux class
A mux class was added which support implementing all the types of muxes needed in COFFE this will make it more easy to update the wire lengths of the muxes and their areas. It will also make it easier to add any muxes of any size to the FPGA architecture later on.